+++ changelog.md
@@ -4,7 +4,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 
 ## [Unreleased]
 
-* ...
+* feat: Scrum Poker session timeout configurable in plugin settings (#19)
 
 ## [3.1.2] - 2018-07-27